Cleaned up the FSMs and made ARM_MOVEMENT have fine tuning.
Split the SHOOT_MID and SUBSTATION_PICKUP into the respective forward and backward versions. Added autonomous functionality into the handle function instead of in execute functions in order to make the FSM simpler. Added an AUTONOMOUS_RETRACT state which is only called in autonomous and is for retracting the telescoping arm all the way to its minimum in auto. Made the motor power being set to the pivot and teleArm motors in the ARM_MOVEMENT state dependent  on the Y and X values of the Joystick to make it easier to fine-tune the positioning of the arm when shooting. The max Motor Power in this state is 0.2, but this will be adjusted after testing on the actual arm.