Dev/5/alex du/spinning intake (#59)
* Updated to represent current fsm

TODO: instead of checking if color is purple, check if yellow, bc it will return not purple if on white side. Also find correct distances that correspond to actual mech.

* Added different lengths for game object. (#47)

In SHOOT_HIGH forward, depending on the game object, changes the target encoder count based on the object.

* Created DistColorTester, which just gives color/dist, and updated distance sensor distances to more closely match hardware values

* Added SmartDashboard code for arm.

* Working Arm Mech Code integrated with grabber.

* Added PID code to the Tele Arm.

* Fixed checkstyle errors

* Updated to represent current fsm

TODO: instead of checking if color is purple, check if yellow, bc it will return not purple if on white side. Also find correct distances that correspond to actual mech.

* Added different lengths for game object. (#47)

In SHOOT_HIGH forward, depending on the game object, changes the target encoder count based on the object.

* Created DistColorTester, which just gives color/dist, and updated distance sensor distances to more closely match hardware values

* Added SmartDashboard code for arm.

* Working Arm Mech Code integrated with grabber.

* Added PID code to the Tele Arm.

* Fixed checkstyle errors

* Removed DistColorTester, because after adding smartdashboard it is no longer needed

* Added standalone grabber testing

Depending on the DI that is plugged into the testboard, the robot will run accordingly.

* Checkstyle

* Fixed testboard

* Fixed Checkstyle

* Changed itemtype from int to more descriptive enum class, and changed armfsm back to using pid

* Updated to represent current fsm

TODO: instead of checking if color is purple, check if yellow, bc it will return not purple if on white side. Also find correct distances that correspond to actual mech.

* Added different lengths for game object. (#47)

In SHOOT_HIGH forward, depending on the game object, changes the target encoder count based on the object.

* Created DistColorTester, which just gives color/dist, and updated distance sensor distances to more closely match hardware values

* Added SmartDashboard code for arm.

* Working Arm Mech Code integrated with grabber.

* Added PID code to the Tele Arm.

* Fixed checkstyle errors

* Added SmartDashboard code for arm.

* Working Arm Mech Code integrated with grabber.

* Added standalone grabber testing

Depending on the DI that is plugged into the testboard, the robot will run accordingly.

* Removed DistColorTester, because after adding smartdashboard it is no longer needed

* Checkstyle

* Fixed testboard

* Fixed Checkstyle

* Added code in both arm and grabber to change the can id for motors depending on if it was a test board or the real robot

* Added code for groundmount, updated hardwaremap and teleop

* Added test board for ground mount and tested

* Added Auto code, may need to test releaseDistance constant in the future

* finished grabber code. wrote ground mount code.

---------

Co-authored-by: Alex <92554149+alexyd88@users.noreply.github.com>
Co-authored-by: Daiji Uchino <103068628+duchino292@users.noreply.github.com>
Co-authored-by: Daiji Uchino <duchino292@student.fuhsd.org>
Co-authored-by: adnx26 <advaithnair26@gmail.com>