Updated all distances, fixed error where color was  incorrect even though it was in the interval
The color sensor has a lower refresh rate than the distance sensor, so the color sensor will report old values even while the distance sensor reads that the object is in the reading interval. Therefore, we only update the item type when we are within the distance interval and the color reported is different than the last color reported, meaning that the color sensor has recently updated.