Arm code with working fine-tuning code.
Adjusted ARM_MOVEMENT state so it uses Joystick Y values for extension/retraction.
Made a fine-tuning button which toggles the fine-tuning state. The inputs for extending/retracting and pivoting are the same but the motor powers are set to a small value.
Made the code pivot and then extend if going from forward to backward or vice versa.