Svr competition (#85)
Updated code for competition with all the subsystems fully integrated. Auto paths have been tested to work on both arm and ground mount, and auto path selection on shuffleboard has been added and is working. CV is also tested for cone and cube pick up as well as tag/alignment. Arm and ground mount pid and encoder values are updated for the new robot configuration.



* added cube and code alignment + pid for apriltag alignment

* Added smarter features.

Added max extension for the arm which changes based on the current angle of the arm. This value is only setting max for horizontal constraints instead of vertical for now. Also added code to make the pivot curve more smooth instead of having a sudden decrease in speed.

* added webcam

* Update build.gradle

* Added code for changing auto path on dashboard.

This code has not been tested yet and currently only functions for auto paths 1-4. Will be updated after rebasing this branch onto main after competition changes have been pushed up.

* Added code for switching ground mount and arm.

Uses the breakout DIO port 0 for changing between ground mount and arm configurations in code.

* updated vs code

* Fixed auto path selection

* new pid values

* Added smarter features.

Added max extension for the arm which changes based on the current angle of the arm. This value is only setting max for horizontal constraints instead of vertical for now. Also added code to make the pivot curve more smooth instead of having a sudden decrease in speed.

* Added code for changing auto path on dashboard.

This code has not been tested yet and currently only functions for auto paths 1-4. Will be updated after rebasing this branch onto main after competition changes have been pushed up.

* Added code for switching ground mount and arm.

Uses the breakout DIO port 0 for changing between ground mount and arm configurations in code.

* updated vs code

* Fixed auto path selection

* Added code for switching ground mount and arm.

The code for switching ground mount and arm is based on whether or not the jumper is in the DIO 0 port on the robot.

* changed print statements

* Implemented known working code

This code does not use the angle for calculating the pivot motor power. The code for getting the motor power to set to based on angle is still in the branch, just not used since it is not fully tested yet.

* Removed unnecessary shuffleboards.

* checkstyle

* Update Constants.java

* resolved merge conflicts + checkstyle

* Fixed robot.java error

* Made the smart dashboard only update when necessary.

Made the String changes on dashboard only happen when the actual values change.

* Constants clean up

* Mode auto path chooser into its own class.

* added PID code for cones and cubes

* updated pipeline indexes

* Fixed missing code

* edits for object alignment + contour selection w/ joystick

* Removed compiler errors.

* Added contour selection feature with buttons

* Added back mech arm button IDs

* Removed duplicate methods

* Added pid code for reflective tape aligning

* testing contour switch

* debugging contour switch

* integrated auto, ground mount, and updated arm code. Tested starting config of the arm and calibrated PID constants. Added throttle to change release speed based on user input.

* added switch contour state

* reset motor power to zero in reset auto and reset tele

* merge cv and mech

* commented out contour selection

* auto changes

* completely tested working autonomus with ground mount :)

* added cv vision state

* auto with arm completely tested, small issues with moving to start state

* pull this daiji

* checkstyle

* autonomous arm fix.

* svr settings for photonvision

* fixed vision state

* removed print statements

* removed cv switch contour

* deleted comments in handleShootForward

* seperated telopinput button values based on joystick

* unziped photonvision-settings file

* renamed photonvision-settings file

---------

Co-authored-by: Tanmayi Dasari <tanmayi.dasari@gmail.com>
Co-authored-by: Daiji Uchino <duchino292@student.fuhsd.org>
Co-authored-by: R0M-GH <rom.malla@outlook.com>
Co-authored-by: adnx26 <advaithnair26@gmail.com>
Co-authored-by: Akshay Anand <95551900+AknA13@users.noreply.github.com>
Co-authored-by: AknA13 <akshaya91306@gmail.com>
Co-authored-by: Daiji Uchino <103068628+duchino292@users.noreply.github.com>
Co-authored-by: Mayand Gulati <mayandg@icloud.com>
Co-authored-by: Tanmayi Dasari <69954364+tanmayi2@users.noreply.github.com>